custom elements

+++ TinyCMS.Data/Nodes/Image.cs
@@ -6,7 +6,7 @@ namespace TinyCMS.Data.Nodes
     public class Image : File
     {
         public override string Type => "image";
-        public int Width { get; set; }
-        public int Height { get; set; }
+        public double Width { get; set; }
+        public double Height { get; set; }
     }
 }
